Create docker-compose.yaml
A configuration running on a host running Ubuntu 22.0 LTS.
Multiple instances of shairport-sync are configured to run on the same host, each sending their audio stream to a separate sound card via ALSA. Docker manages the virtual network and IP address of each instance of shairport-sync on the virtual network. Care must be taken to not overlap with the primary DHCP server.